ai-notes-summarizer/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
transformers
torch
boto3
python-dotenv
pydantic
python-multipart
pypdf2